requirements.txt
fastapi
uvicorn[standard]
PyYAML
aiofiles
python-multipart